Capitalizes the first letter of every word

Capitalizes the first letter of every word

capitalize_words(words, lowercase_of = TRUE)

Arguments

words
A string or vector of strings with words you want capitalized

lowercase_of
If TRUE (default), keeps the string " of " to be lowercased as is custom in English writing (e.g. District of Columbia).

Value

The original string with the first letter of each word capitalized